Add separate storage layer
The overall program organization is now relatively stable.
The CLI command code calls into the parameter parsing and the storage
system, using both to execute the instruction given in parameters.

The main expected change in the future is the addition of a
configuration into the storage system, to keep track of a default
database file, default user, and various other settings.