Test the `user list` command
It requires improving the functional tests with a before() function to
create a per-test database, avoiding cross-test pollution.
We take advantage of this necessity to also add support for a
$EXPENSE_DB environment variable throughout all commands, which allows
anyone to use a non-default database.

We take the opportunity to also improve error reporting in functional
tests, and to allow arbitrary stdout/stderr checks instead of just exact
string matches.

The help messages also now include the description of the command being
queried (a past oversight).